+++ src/core/modules/project_management/infrastructure/reporting/models/report_models.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 
 from src.core.modules.project_management.application.resources.resource_load_engine import (
     ResourceUtilizationBand,
@@ -100,15 +101,15 @@ class BaselineComparisonRow:
     baseline_a_start: date | None
     baseline_a_finish: date | None
     baseline_a_duration_days: int | None
-    baseline_a_planned_cost: float | None
+    baseline_a_planned_cost: Decimal | None
     baseline_b_start: date | None
     baseline_b_finish: date | None
     baseline_b_duration_days: int | None
-    baseline_b_planned_cost: float | None
+    baseline_b_planned_cost: Decimal | None
     start_shift_days: int | None
     finish_shift_days: int | None
     duration_delta_days: int | None
-    planned_cost_delta: float
+    planned_cost_delta: Decimal
     change_type: str
 
 @dataclass
